+++ src/DependencyResolutionException.php
@@ -5,7 +5,7 @@
  * www.bueller.ca/di
  *
  * DependencyResolutionException.php
- * @copyright Copyright (c) 2016 Matt Ferris
+ * @copyright Copyright (c) 2023 Matt Ferris
  * @author Matt Ferris <matt@bueller.ca>
  *
  * Licensed under BSD 2-clause license
@@ -15,8 +15,10 @@
 namespace MattFerris\Di;
 
 use Exception;
+use Psr\Container\ContainerExceptionInterface;
 
-class DependencyResolutionException extends Exception
+
+class DependencyResolutionException extends Exception implements ContainerExceptionInterface
 {
     /**
      * @var string The definition type
@@ -27,8 +29,7 @@ class DependencyResolutionException extends Exception
      * @param string $type The definition type
      * @param string $key The definition key
      */
-    public function __construct($type)
-    {
+    public function __construct($type) {
         $this->type = $type;
         $msg = 'Failed to resolve dependency "'.$type.'"';
         parent::__construct($msg);
@@ -37,8 +38,7 @@ class DependencyResolutionException extends Exception
     /**
      * @return string The definition type
      */
-    public function getType()
-    {
+    public function getType(): string {
         return $this->type;
     }
 }