+++ src/ServiceProvider.php
@@ -5,7 +5,7 @@
  * www.bueller.ca/di
  *
  * ServiceProvider.php
- * @copyright Copyright (c) 2015 Matt Ferris
+ * @copyright Copyright (c) 2023 Matt Ferris
  * @author Matt Ferris <matt@bueller.ca>
  *
  * Licensed under BSD 2-clause license
@@ -17,13 +17,13 @@ namespace MattFerris\Di;
 use MattFerris\Provider\ProviderInterface;
 use MattFerris\Provider\InvalidConsumerException;
 
+
 class ServiceProvider implements ProviderInterface
 {
     /**
      * {@inheritDoc}
      */
-    public function provides($consumer)
-    {
+    public function provides($consumer) {
         if (!($consumer instanceof ContainerInterface)) {
             throw new InvalidConsumerException();
         }